Split out matplotlib and qcstyle as separate modules (#1105)
This commit starts the decomposition of the _circuit_visualization
module into having separate modules per output backend. The
_circuit_visualization module has is quite bulky containing the
implementation of 2 different visualization backends. This starts by
splitting out matplotlib and QCStyle. The latex backend will be done
soon, but it currently in flux with multiple open pull requests. Once
development calms on that backend it too will be spit into a separate
module.

Starts #1104
