+++ qiskit/tools/visualization/_qcstyle.py
@@ -1,0 +1,104 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree
+
+# pylint: disable=invalid-name,anomalous-backslash-in-string,missing-docstring
+
+
+class QCStyle:
+    def __init__(self):
+        self.tc = '#000000'
+        self.sc = '#000000'
+        self.lc = '#000000'
+        self.cc = '#778899'
+        self.gc = '#ffffff'
+        self.gt = '#000000'
+        self.bc = '#bdbdbd'
+        self.bg = '#ffffff'
+        self.fs = 13
+        self.sfs = 8
+        self.disptex = {
+            'id': 'id',
+            'u0': 'U_0',
+            'u1': 'U_1',
+            'u2': 'U_2',
+            'u3': 'U_3',
+            'x': 'X',
+            'y': 'Y',
+            'z': 'Z',
+            'h': 'H',
+            's': 'S',
+            'sdg': 'S^\\dagger',
+            't': 'T',
+            'tdg': 'T^\\dagger',
+            'rx': 'R_x',
+            'ry': 'R_y',
+            'rz': 'R_z',
+            'reset': '\\left|0\\right\\rangle'
+        }
+        self.dispcol = {
+            'id': '#ffffff',
+            'u0': '#ffffff',
+            'u1': '#ffffff',
+            'u2': '#ffffff',
+            'u3': '#ffffff',
+            'x': '#ffffff',
+            'y': '#ffffff',
+            'z': '#ffffff',
+            'h': '#ffffff',
+            's': '#ffffff',
+            'sdg': '#ffffff',
+            't': '#ffffff',
+            'tdg': '#ffffff',
+            'rx': '#ffffff',
+            'ry': '#ffffff',
+            'rz': '#ffffff',
+            'reset': '#ffffff',
+            'target': '#ffffff',
+            'meas': '#ffffff'
+        }
+        self.latexmode = True
+        self.pimode = False
+        self.fold = 20
+        self.bundle = False
+        self.barrier = False
+        self.index = False
+        self.compress = True
+        self.figwidth = -1
+        self.dpi = 150
+        self.margin = [2.0, 0.0, 0.0, 0.3]
+        self.cline = 'doublet'
+        self.reverse = False
+
+    def set_style(self, dic):
+        self.tc = dic.get('textcolor', self.tc)
+        self.sc = dic.get('subtextcolor', self.sc)
+        self.lc = dic.get('linecolor', self.lc)
+        self.cc = dic.get('creglinecolor', self.cc)
+        self.gt = dic.get('gatetextcolor', self.tc)
+        self.gc = dic.get('gatefacecolor', self.gc)
+        self.bc = dic.get('barrierfacecolor', self.bc)
+        self.bg = dic.get('backgroundcolor', self.bg)
+        self.fs = dic.get('fontsize', self.fs)
+        self.sfs = dic.get('subfontsize', self.sfs)
+        self.disptex = dic.get('displaytext', self.disptex)
+        for key in self.dispcol.keys():
+            self.dispcol[key] = self.gc
+        self.dispcol = dic.get('displaycolor', self.dispcol)
+        self.latexmode = dic.get('latexdrawerstyle', self.latexmode)
+        self.pimode = dic.get('usepiformat', self.pimode)
+        self.fold = dic.get('fold', self.fold)
+        if self.fold < 2:
+            self.fold = -1
+        self.bundle = dic.get('cregbundle', self.bundle)
+        self.barrier = dic.get('plotbarrier', self.barrier)
+        self.index = dic.get('showindex', self.index)
+        self.compress = dic.get('compress', self.compress)
+        self.figwidth = dic.get('figwidth', self.figwidth)
+        self.dpi = dic.get('dpi', self.dpi)
+        self.margin = dic.get('margin', self.margin)
+        self.cline = dic.get('creglinestyle', self.cline)
+        self.reverse = dic.get('reversebits', self.reverse)
