+++ qiskit/tools/visualization/_utils.py
@@ -1,0 +1,21 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree
+
+"""Common visualization utilties."""
+
+import PIL
+
+
+def _trim(image):
+    """Trim a PIL image and remove white space."""
+    background = PIL.Image.new(image.mode, image.size, image.getpixel((0, 0)))
+    diff = PIL.ImageChops.difference(image, background)
+    diff = PIL.ImageChops.add(diff, diff, 2.0, -100)
+    bbox = diff.getbbox()
+    if bbox:
+        image = image.crop(bbox)
+    return image
